+++ scripts/macros/monsterFeatures/ghast/ghast.js
@@ -1,0 +1,4 @@
+import {stench} from '../troglodyte/stench.js';
+export let ghast = {
+    'stench': stench
+}