+++ scripts/macros/monsterFeatures/hezrou/hezrou.js
@@ -1,0 +1,4 @@
+import {stench} from '../troglodyte/stench.js';
+export let hezrou = {
+    'stench': stench
+}