+++ scripts/macros/monsterFeatures/leucrotta/leucrotta.js
@@ -1,0 +1,4 @@
+import {stench} from '../troglodyte/stench.js';
+export let leucrotta = {
+    'stench': stench
+}