+++ scripts/macros/monsterFeatures/shadowghast/shadowghast.js
@@ -1,0 +1,4 @@
+import {stench} from '../troglodyte/stench.js';
+export let shadowghast = {
+    'stench': stench
+}