+++ scripts/macros/monsterFeatures/stenchKow/stenchKow.js
@@ -1,0 +1,4 @@
+import {stench} from '../troglodyte/stench.js';
+export let stenchKow = {
+    'stench': stench
+}