+++ scripts/macros/monsterFeatures/troglodyte/troglodyte.js
@@ -1,0 +1,4 @@
+import {stench} from './stench.js';
+export let troglodyte = {
+    'stench': stench
+}